refactor: move GeometryField into its own module
Extract the GeometryField registry out of constants.py -- whose legacy
"constants for TIM" header made it read as mission-specific -- into a
dedicated geometry_fields.py. constants.py keeps a backward-compatible
re-export and its module docstring now describes its actual generic
contents; geometry.py imports from the new module.
